+++ FoodSwipe/App.js
@@ -1,21 +1,15 @@
 import React, { Component } from 'react';
-import { View, StyleSheet } from 'react-native';
+import { View } from 'react-native';
 import { createStore } from 'redux'
 import { Provider } from 'react-redux'
 import { NativeRouter, Route } from 'react-router-native'
 
 import reducer from './reducers'
 
-// import FavoritesList from './containers/FavoritesList'
 import Header from './components/Header'
 import FilterPage from './components/FilterPage'
-import SwipeCards from './containers/SwipeCards'
-
-const Cards = () => (
-	<View style={styles.cards}>
-		<SwipeCards />
-	</View>
-);
+import CardsContainer from './containers/CardsContainer'
+// import FavoritesList from './containers/FavoritesList'
 
 export default class App extends Component{
 	// const store = createStore(reducer)
@@ -25,10 +19,10 @@ export default class App extends Component{
     return (
 
 				<NativeRouter>
-					<View style={styles.containers}>
+					<View style={{flex: 1}}>
 						<Header headerText={"Food Swipe"}/>
 						<Route exact path="/" component={FilterPage}/>
-						<Route path="/cards" component={Cards}/>
+						<Route path="/cards" component={CardsContainer}/>
 					</View>
 				</NativeRouter>
 
@@ -36,15 +30,3 @@ export default class App extends Component{
   }
 }
 // TODO: <Route path="/favorites" component={FavoritesList}/>
-
-const styles = StyleSheet.create({
-	containers: {
-		flex: 1
-	},
-	cards: {
-		flex: 1,
-		justifyContent: 'center',
-		alignItems: 'center',
-		backgroundColor: '#F5FCFF',
-	}
-});
